--- .npmignore
@@ -1,20 +1,0 @@
-lib-cov
-*.seed
-*.log
-*.csv
-*.dat
-*.out
-*.pid
-*.gz
-
-pids
-logs
-results
-
-npm-debug.log
-node_modules
-coverage
-examples
-test
-assets
-.idea
