+++ test/loader.test.js
@@ -84,47 +84,7 @@ describe('loader', () => {
       `${GET_URL_CODE}module.exports = "<h3>#{number} {customer}</h3>\\n<p>   {title}   </p>";`
     );
   });
-  it('should minimize', () => {
-    const result = loader.call(
-      {
-        query: '?minimize',
-      },
-      '<!-- comment --><h3 customAttr="">#{number} {customer}</h3>\n<p>   {title}   </p>\n\t <!-- comment --> <img src="image.png" />'
-    );
 
-    expect(result).toBe(
-      `${GET_URL_CODE}module.exports = "<h3 customattr=\\"\\">#{number} {customer}</h3> <p> {title} </p> <img src=" + __url__(require("./image.png")) + " />";`
-    );
-  });
-  // https://github.com/webpack/webpack/issues/752
-  it('should not remove attributes by default', () => {
-    const result = loader.call(
-      {
-        query: '?minimize',
-      },
-      '<input type="text" />'
-    );
-
-    expect(result).toBe(
-      `${GET_URL_CODE}module.exports = "<input type=text />";`
-    );
-  });
-  it('should preserve comments', () => {
-    const result = loader.call(
-      {
-        query: {
-          minimize: {
-            removeComments: false,
-          },
-        },
-      },
-      '<!-- comment --><h3 customAttr="">#{number} {customer}</h3><p>{title}</p><!-- comment --><img src="image.png" />'
-    );
-
-    expect(result).toBe(
-      `${GET_URL_CODE}module.exports = "<!-- comment --><h3 customattr=\\"\\">#{number} {customer}</h3><p>{title}</p><!-- comment --><img src=" + __url__(require("./image.png")) + " />";`
-    );
-  });
   it('should preserve escaped quotes', () => {
     const result = loader.call(
       {},
@@ -136,58 +96,6 @@ describe('loader', () => {
     );
   });
 
-  it('should preserve comments and white spaces when minimizing (via webpack config property)', () => {
-    const result = loader.call(
-      {
-        query: {
-          minimize: {
-            removeComments: false,
-            collapseWhitespace: false,
-          },
-        },
-      },
-      '<!-- comment --><h3 customAttr="">#{number} {customer}</h3><p>{title}</p>    <!-- comment -->    <img src="image.png" />'
-    );
-
-    expect(result).toBe(
-      `${GET_URL_CODE}module.exports = "<!-- comment --><h3 customattr=\\"\\">#{number} {customer}</h3><p>{title}</p>    <!-- comment -->    <img src=" + __url__(require("./image.png")) + " />";`
-    );
-  });
-
-  it('should preserve comments and white spaces when minizing (via webpack config property)', () => {
-    const result = loader.call(
-      {
-        query: {
-          minimize: {
-            removeComments: false,
-            collapseWhitespace: false,
-          },
-        },
-      },
-      '<!-- comment --><h3 customAttr="">#{number} {customer}</h3><p>{title}</p>    <!-- comment -->    <img src="image.png" />'
-    );
-
-    expect(result).toBe(
-      `${GET_URL_CODE}module.exports = "<!-- comment --><h3 customattr=\\"\\">#{number} {customer}</h3><p>{title}</p>    <!-- comment -->    <img src=" + __url__(require("./image.png")) + " />";`
-    );
-  });
-
-  it('should treat attributes as case sensitive', () => {
-    const result = loader.call(
-      {
-        query: {
-          minimize: {
-            caseSensitive: true,
-          },
-        },
-      },
-      '<!-- comment --><h3 customAttr="">#{number} {customer}</h3><p>{title}</p><!-- comment --><img src="image.png" />'
-    );
-
-    expect(result).toBe(
-      `${GET_URL_CODE}module.exports = "<h3 customAttr=\\"\\">#{number} {customer}</h3><p>{title}</p><img src=" + __url__(require("./image.png")) + " />";`
-    );
-  });
   it('should not translate root-relative urls (without root query)', () => {
     const result = loader.call({}, 'Text <img src="/image.png">');
